+++ storage/albums.json
@@ -1,7 +1,1 @@
-[
-  {"id": 878, "on_spotify": false, "publish_date": "16/10/1996"},
-  {"id": 611, "on_spotify": false, "publish_date": "16/11/02"},
-  {"id": 177, "on_spotify": false, "publish_date": "16/10/1996"},
-  {"id": 887, "on_spotify": false, "publish_date": "13/02/2022"},
-  {"id": 13, "on_spotify": false, "publish_date": "23/03/2022"}
-]
+[{"id":388,"on_spotify":false,"publish_date":"16/10/1996"},{"id":900,"on_spotify":false,"publish_date":"16/11/02"}]
