+++ storage/books.json
@@ -1,1 +1,1 @@
-[{"id":85,"title":"Harry potter","publisher":"Men","publish_date":"02/02/2001","cover_state":"Good"},{"id":466,"title":"LOTR","publisher":"Tolkien","publish_date":"03/03/1990","cover_state":"Bad"}]
+[{"id":293,"title":"Harry potter","publisher":"Men","publish_date":"02/02/2001","cover_state":"Good"},{"id":147,"title":"LOTR","publisher":"Tolkien","publish_date":"03/03/1990","cover_state":"Bad"}]
