+++ storage/games.json
@@ -1,1 +1,1 @@
-[{"id":794,"multiplayer":"true","publish_date":"12/02/2015"}]
+[{"id":706,"multiplayer":548,"publish_date":"12/02/2015"},{"id":767,"multiplayer":863,"publish_date":"17/10/1998"}]
