+++ storage/genres.json
@@ -1,6 +1,1 @@
-[
-  {"id": 249, "name": "Rock"},
-  {"id": 672, "name": "Pop"},
-  {"id": 151, "name": "Pop alt"},
-  {"id": 880, "name": "love"}
-]
+[{"id":855,"name":"Rock"},{"id":852,"name":"Pop"},{"id":26,"name":"Pop alt"}]
